+++ backend/app/routes/notifications.py
@@ -1,6 +1,6 @@
 from . import api
-from app.utils.validators import require_auth
-from app.utils.response import success_response, error_response, warning_response
+from app.utils.auth import require_auth
+from app.utils.responses import success_response, error_response, warning_response
 from app.services.user import fetch_user
 from app.services.calendar import fetch_medicine_name
 from app.db.connection import get_connection
