+++ backend/app/routes/shared_users_medicines.py
@@ -1,11 +1,11 @@
 from flask import request, g
-from app.utils.validators import require_auth
+from app.utils.auth import require_auth
 from . import api
-from app.utils.response import success_response, error_response, warning_response
-from app.services.verifications import verify_calendar_share
-from app.services.medicines import get_boxes, update_box, create_box, delete_box, restock_box
+from app.utils.responses import success_response, error_response, warning_response
+from app.services.calendar import verify_calendar_share
+from app.services.medication import get_boxes, update_box, create_box, delete_box, restock_box
 import time
-from app.services.pillulier import use_pillulier
+from app.services.medication import use_pillulier
 from datetime import datetime, timezone
 
 ERROR_CALENDAR_NOT_FOUND = "calendrier non trouvé"
