+++ backend/app/routes/tokens.py
@@ -1,12 +1,12 @@
-from app.utils.response import success_response, error_response, warning_response
-from app.utils.validators import require_auth
+from app.utils.responses import success_response, error_response, warning_response
+from app.utils.auth import require_auth
 from datetime import datetime, timezone
 from . import api
 import time
 from flask import request, g
 from app.db.connection import get_connection
 from app.services.calendar import generate_calendar_schedule
-from app.services.verifications import verify_calendar, verify_token_owner, verify_token
+from app.services.calendar import verify_calendar, verify_token_owner, verify_token
 
 ERROR_UNAUTHORIZED_ACCESS = "accès refusé"
