+++ backend/app/services/__init__.py
@@ -1,8 +1,5 @@
 from .calendar import *
-from .medicines import *
-from .user import *
-from .pdf import *
+from .medication import *
 from .notifications import *
-from .pillulier import *
-from .verifications import *
-from .stock import *
+from .documents import *
+from .user import *
