+++ backend/app/services/calendar/__init__.py
@@ -1,0 +1,2 @@
+from .core import *
+from .verification import *
