+++ backend/app/services/documents/pdf.py
@@ -2,7 +2,7 @@ from io import BytesIO
 from reportlab.lib.pagesizes import A4
 from reportlab.lib.styles import getSampleStyleSheet
 from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer
-from app.services.medicines import get_medicines_for_calendar
+from app.services.medication import get_medicines_for_calendar
 
 moment_map = {
     "morning": "matin",
