+++ backend/app/services/medication/__init__.py
@@ -1,0 +1,3 @@
+from .boxes import *
+from .pillbox import *
+from .stock import *
