+++ backend/app/services/medication/pillbox.py
@@ -1,5 +1,5 @@
 from app.db.connection import get_connection
-from app.services.stock import process_box_decrement, check_low_stock_and_notify_for_calendar
+from app.services.medication import process_box_decrement, check_low_stock_and_notify_for_calendar
 
 def use_pillulier(calendar_id, start_date):
     """
