+++ backend/app/services/notifications/__init__.py
@@ -1,0 +1,2 @@
+from .core import *
+from .messaging import *
