+++ backend/app/services/notifications/core.py
@@ -5,7 +5,7 @@ from app.db.connection import get_connection
 from app.services.calendar import fetch_calendar, fetch_medicine_name
 from app.services.messaging import send_email, send_sms
 from app.services.user import fetch_user
-from app.utils.logger import log_backend
+from app.utils.logging import log_backend
 from app.config import Config
 import traceback
